move all remaining used queries to new db
with this commit all the remaining queries in use have been moved
to the new db abstraction.

this means that running `zig build test` will now check that all
queries are semantically correct for sqlite, and schema changes will
produce compile errors everywhere in the code where the same query
is being used

+++ src/awebo/User.zig
@@ -9,7 +9,7 @@ const context = @import("options").context;
 // - u4 max 16 clients connected
 // - u2 max 4 media streams (cam, voice, screenshare, screenshare audio)
 
-pub const Id = u32;
+pub const Id = u64;
 
 id: Id,
 created: u32 = 0,
